### 2018-10-04 #996 @whedon
[PRE REVIEW]: iRF: Extracting interactions from random forests
**Submitting author:** @karlkumbier (<a href="http://orcid.org/0000-0001-6521-1173">Karl Kumbier</a>)
**Repository:** <a href="https://github.com/sumbose/iRF" target ="_blank">https://github.com/sumbose/iRF</a>
**Version:** v2.0.0
