### 2018-10-05 ec7d44a
edits to paper
