+++ paper.md
@@ -6,9 +6,9 @@ tags:
   - Interpretable machine learning
 authors:
   - name: Sumanta Basu
-    affiliation: "\*, 1, 2"
+    affiliation: "*, 1, 2"
   - name: Karl Kumbier
-    affiliation: "\*, 3"
+    affiliation: "*, 3"
   - name: James B. Brown
     affiliation: "3, 4, 5"
   - name: Bin Yu
