### 2018-10-05 37975f4
edits to paper
