### 2018-10-05 8bdb125
edits to paper
