### 2018-10-05 dc67096
edits to paper
