+++ paper.md
@@ -14,6 +14,8 @@ authors:
   - name: Bin Yu
     affiliation: "c, f"
 affiliations:
+  - name: Denotes equal contribution
+   index: 1;
  - name: Department of Biological Statistics and Computational Biology, Cornell University
    index: a
  - name: Department of Statistical Science, Cornell University
@@ -26,8 +28,6 @@ affiliations:
    index: e
  - name: Department of Electrical Engineering and Computer Sciences, University of California, Berkeley 
    index: f
- - name: Denotes equal contribution
-   index: 1;
 date: 2 October 2018
 bibliography: paper.bib
 ---
