### 2018-10-05 476f2ed
edits to paper
