### 2018-10-12 #996 @arfon
@arokem - would you be willing to edit this submission for JOSS?
