+++ R/readForest.R
@@ -1,136 +1,117 @@
-readForest <- function(rfobj, x, y=NULL, 
-                       return.node.feature=TRUE,
-                       wt.pred.accuracy=FALSE,
+readForest <- function(rfobj, x, 
+                       return.node.feature=TRUE, 
+                       return.node.obs=FALSE,
+                       varnames.grp=NULL,
+                       get.split=FALSE,
+                       first.split=TRUE,
                        n.core=1){
- 
+  
   if (is.null(rfobj$forest))
     stop('No Forest component in the randomForest object')
-  if (wt.pred.accuracy & is.null(y))
-    stop('y required to evaluate prediction accuracy')
-  if (n.core > 1) registerDoParallel(cores=n.core) 
+  if (is.null(varnames.grp)) varnames.grp <- 1:ncol(x)
+ 
   ntree <- rfobj$ntree
-  p <- ncol(x)
   n <- nrow(x)
+  p <- length(unique(varnames.grp))
   out <- list()
   
-  # read leaf node data from each tree in the forest
-  tt <- NULL
-  rd.forest <- foreach(tt=1:ntree) %dopar% {
-    readTree(rfobj=rfobj, k=tt, x=x, y=y,
-      return.node.feature=return.node.feature,
-      wt.pred.accuracy=wt.pred.accuracy)
-  }
- 
+  # Determine leaf nodes for observations in x
+  prf <- predict(rfobj, newdata=x, nodes=TRUE)
+  nodes <- attr(prf, 'nodes')
+  
+  # read leaf node data from each tree in the forest 
+  rd.forest <- mclapply(1:ntree, readTree, rfobj=rfobj, x=x,
+                        nodes=nodes,
+                        varnames.grp=varnames.grp,
+                        return.node.feature=return.node.feature,
+                        return.node.obs=return.node.obs,
+                        get.split=get.split,
+                        first.split=first.split,
+                        mc.cores=n.core)
+  
+  # aggregate node level metadata across forest
   out$tree.info <- rbindlist(lapply(rd.forest, function(tt) tt$tree.info))
-  # aggregate sparse feature matrix across forest
+  
+  # aggregate sparse node level feature matrix across forest
   nf <- lapply(rd.forest, function(tt) tt$node.feature)
   nf <- aggregateNodeFeature(nf)
-  out$node.feature <- sparseMatrix(i=nf[,1], j=nf[,2], dims=c(max(nf[,1]), p))
+  
+  if (get.split) 
+    out$node.feature <- sparseMatrix(i=nf[,1], j=nf[,2], x=nf[,3], 
+                                     dims=c(max(nf[,1]), 2 * p))
+  else
+    out$node.feature <- sparseMatrix(i=nf[,1], j=nf[,2],
+                                     dims=c(max(nf[,1]), 2 * p))
+
+ 
+  # aggregate sparse node level observation matrix across forest
+  if (return.node.obs) {
+    nobs <- lapply(rd.forest, function(tt) tt$node.obs)
+    nobs <- aggregateNodeFeature(nobs)
+    out$node.obs <- sparseMatrix(i=nobs[,1], j=nobs[,2], dims=c(max(nf[,1]), n))
+  }
+
   return(out)
   
 }
 
-readTree <- function(rfobj, k, x, y, return.node.feature, wt.pred.accuracy) {
-  n <- nrow(x)
-  p <- ncol(x)
-  ntree <- rfobj$ntree
+readTree <- function(rfobj, k, x, nodes,
+                     varnames.grp=1:ncol(x), 
+                     return.node.feature=TRUE,
+                     return.node.obs=FALSE,
+                     get.split=FALSE,
[…]
