### 2018-11-07 2c9506a
updates for signed iRF
