+++ R/predict.randomForest.R
@@ -184,6 +184,7 @@
                   proximity = as.double(proxmatrix),
                   nodes = as.integer(nodes),
                   nodexts = as.integer(nodexts),
+                  DUP=FALSE,
                   PACKAGE = "iRF")[keepIndex]
             ## Apply bias correction if needed.
             yhat <- rep(NA, length(rn))
@@ -241,6 +242,7 @@
                  prox = as.integer(proximity),
                  proxmatrix = as.double(proxmatrix),
                  nodes = as.integer(nodes),
+                 DUP=FALSE,
                  PACKAGE = "iRF")
         if (out.type > 1) {
             out.class.votes <- t(matrix(t1$countts, nrow = nclass, ncol = ntest))
