+++ R/iRF.R
@@ -4,258 +4,251 @@ iRF <- function(x, y,
                 n.iter=5, 
                 ntree=500, 
                 n.core=1, 
-                mtry.select.prob=rep(1/ncol(x), ncol(x)), 
-                keep.impvar.quantile=NULL, 
+                mtry.select.prob=rep(1, ncol(x)),
                 interactions.return=NULL, 
-                wt.pred.accuracy=FALSE, 
-                cutoff.unimp.feature=0,  
-                rit.param=list(depth=5, ntree=100, nchild=2, class.id=1, class.cut=NULL), 
-                varnames.grp=NULL, 
-                n.bootstrap=30,
-                bootstrap.forest=TRUE, 
+                rit.param=list(depth=5, ntree=500, 
+                               nchild=2, class.id=1, 
+                               min.nd=1, class.cut=NULL), 
+                varnames.grp=colnames(x), 
+                n.bootstrap=1,
+                select.iter=FALSE,
+                signed=TRUE,
                 verbose=TRUE,
-               ...) {
+                ...) {
  
-
-  if (!is.matrix(x) | (!is.null(xtest) & !is.matrix(xtest)))
-    stop('either x or xtest is not a matrix !')
-  if (!is.numeric(x) | (!is.null(xtest) & !is.numeric(xtest)))
-    stop('either x or xtest is not a numeric matrix!')
+  # Check inputs
   if (ncol(x) < 2 & !is.null(interactions.return))
     stop('cannot find interaction - X has less than two columns!')
   if (any(interactions.return > n.iter))
-    stop('interaction iteration to return greater than n.iter')
+    stop('selected iteration greater than n.iter')
+  if (length(mtry.select.prob) != ncol(x))
+    stop('length mtry.select.prop must equal number of features')
+  if (!is.null(xtest) & ncol(xtest) != ncol(x))
+    stop('training/test data must have same number of features')
+  if (!is.null(xtest) & is.null(ytest))
+    stop('test set responses not indicated')
+
+  # Check all RIT and set to defaul if missing
+  if (is.null(rit.param$depth)) rit.param$depth <- 5
+  if (is.null(rit.param$ntree)) rit.param$ntree <- 500
+  if (is.null(rit.param$nchild)) rit.param$nchild <- 2
+  if (is.null(rit.param$class.id) & is.factor(y)) rit.param$class.id <- 1
+  if (is.null(rit.param$min.nd)) rit.param$min.nd <- 1
+  if (is.null(rit.param$class.cut) & is.numeric(y)) 
+    rit.param$class.cut <- median(y)
  
   n <- nrow(x)
   p <- ncol(x)
   class.irf <- is.factor(y)
-  if (n.core > 1) registerDoParallel(cores=n.core)  
-
+  importance <- ifelse(class.irf, 'MeanDecreaseGini', 'IncNodePurity')
+   
   rf.list <- list()
-  if (!is.null(interactions.return)) {
-    interact.list <- list()
+  if (!is.null(interactions.return) | select.iter) {
     stability.score <- list()
-    prevalence <- list()
+    importance.score <- list()
   }
-  
+
   # Set number of trees to grow in each core
   a <- floor(ntree / n.core) 
   b <- ntree %% n.core
   ntree.id <- c(rep(a + 1, b), rep(a, n.core - b))
   
+  registerDoMC(n.core)
   for (iter in 1:n.iter) {
     
-    ## 1: Grow Random Forest on full data
+    # Grow Random Forest on full data
     print(paste('iteration = ', iter))
     rf.list[[iter]] <- foreach(i=1:length(ntree.id), .combine=combine, 
-                               .multicombine=TRUE, .packages='iRF') %dopar% {
+                               .multicombine=TRUE, .packages='iRF') %dorng% {
[…]
