+++ R/MDSplot.R
@@ -9,7 +9,7 @@ MDSplot <- function(rf, fac, k=2, palette=NULL, pch=20, ...) {
     colnames(rf.mds$points) <- paste("Dim", 1:k)
     nlevs <- nlevels(fac)
     if (is.null(palette)) {
-        palette <- if (nlevs < 12)
+        palette <- if (require(RColorBrewer) && nlevs < 12)
             brewer.pal(nlevs, "Set1") else rainbow(nlevs)
     }
     if (k <= 2) {
