+++ man/readForest.Rd
@@ -1,29 +1,39 @@
 \name{readForest}
 \alias{readForest}
 \title{Pass data through a fitted forest, record node characteristics [works
-  for binary classification with continuous variables]}
+  for continuous variables]}
 \description{
-  Passes a feature matrix (and optionally a label vector) through a fitted random forest
-  object, records size (and Gini impurity) of each node. Optionally,
-  for every node, returns the features used to define the rule and the data
-  points falling in that node. Uses \code{mclapply} function to
-  distribute computation across available cores.
+  Passes a feature matrix (and optionally a response vector) through a fitted
+  random forest object, records size of each node. Optionally,for every node,
+  returns the features used to define the rule and the data points falling in
+  that node. Uses \code{foreach} function to distribute computation across
+  available cores.
 }
 \usage{
-  readForest(rfobj, x, y=NULL, return.node.feature=TRUE, 
-    wt.pred.accuracy=FALSE, n.core=1)
+  function(rfobj, x, 
+           return.node.feature=TRUE,
+           return.node.obs=FALSE, 
+           varnames.group=NULL,
+           get.split=FALSE,
+           first.split=TRUE,
+           n.core=1)
 }
 \arguments{
   \item{rfobj}{a fitted \code{randomForest} object with the
     \code{forest} component in it}
   \item{x}{numeric matrix with the same number of predictors used in
     \code{rfobj} fit}
-  \item{y}{a numeric vector specifying response values}
-  \item{return.node.feature}{if TRUE, returns a matrix containing features
-    used to define the decision rule associated with a node}
-  \item{wt.pred.accuracy}{Should leaf nodes be sampled proportional to both
-    size and decrease in variabiliy of responses?}
-  \item{n.core}{number of cores across which tree reading will be distributed}
+  \item{return.node.feature}{if TRUE, returns a sparse matrix that encodes the
+    features and inequality directions used to define the decision rule associated 
+    with a node}
+  \item{return.node.obs}{if TRUE, returns a sparse matrix indicating the
+    observations that fall in a leaf node}
+  \item{varnames.grp}{a vector of length ncol(x) indicating feature groupings}
+  \item{get.split}{if TRUE, entries of \code{node.feature} indicate the
+    threshold associated with a feature split}
+  \item{first.split}{if TRUE, only the threshold associated with the first time
+    a feature is selected will be recorded}
+  \item{n.core}{the number of cores to parallelize over}
 }
 \value{ A list containing the following items:
   \item{tree.info}{a data frame with number of rows equal to total number of
@@ -39,6 +49,10 @@
     with ncol(x) columns, each row corresponding to a leaf node in the
     forest. The entries indicate which features were used to define the
     decision rule associated with a node}
+  \item{node.feature}{if return.node.obs = TRUE, returns a sparse matrix
+    with nrow(x) columns, each row corresponding to a leaf node in the
+    forest. The entries indicate which observations fall in the corresponding
+    leaf node.}
 }
 
 \seealso{
@@ -46,7 +60,11 @@
 }
 
 \examples{
-  n = 50; p = 10
+  library(doMC)
+  registerDoMC()
+  options(cores = 7)
+
+  n = 200; p = 250
   X = array(rnorm(n*p), c(n, p))
   Y = (X[,1]>0.35 & X[,2]>0.35)|(X[,5]>0.35 & X[,7]>0.35)
[…]

[…]
