### 2018-11-07 5e66411
debug for null test sets
