### 2018-11-11 #996 @road2stat
@karlkumbier @arokem -sure thing! I'm happy to review this paper. Just let me know what I need to do next.
