### 2018-11-12 #996 @arokem
@road2stat : please take a look at the guidelines for reviewers here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
