### 2018-11-12 #996 @arokem
@road2stat : please head over to #1077 for your reviewer checklist. Please let me know if you have any questions, and feel free to tag me on review issues that you put up on the software repo. I'll go ahead and close this issue.
