### 2018-11-12 sumbose/iRF#12 @road2stat
[JOSS review] Fix affiliation format in paper.md
It seems the format of two affiliations in `paper.md`is not valid YAML... should be in a single line. That's probably why whedon had troubles generating pdf.

To fix this, please merge line 25 and 26; also line 28 and 29, so the paper can be compiled properly.

https://github.com/openjournals/joss-reviews/issues/1077
