### 2018-11-12 sumbose/iRF#13 @road2stat
[JOSS review] Installation guideline improvements
Some items in the readme on installation could be improved:

1. It looks like the package is already available from CRAN (awesome!), so it is usually good practice to include the installation guidelines for the CRAN version in addition to the GitHub version. A simple example: https://github.com/road2stat/liftr#installation

2. I would also suggest removing the "Alternatively, ... R CMD INSTALL" method because it is more manual and not too different from using `devtools::install_github`, at least for regular users who don't want to modify the source code before using the software.

3. gfortran. I'd probably mention that to compile iRF from source, it needs gfortran, and point to the official CRAN page https://cran.r-project.org/bin/macosx/tools/ for users who want to install it under MacOS. I would also remove the command line since it doesn't handle the complexity of system versions and gfortran versions.

https://github.com/openjournals/joss-reviews/issues/1077
