### 2018-11-12 sumbose/iRF#14 @road2stat
[JOSS review] Make vignette available online
Some documentation improvements:

1. It is often very useful to make vignette directly readable online. The fastest way I can think of is using [pkgdown](https://pkgdown.r-lib.org/) and host the website with GitHub Pages.

2. When it's ready, please add the website or vignette link (if you use GitHub Pages, it would be https://sumbose.github.io/iRF/ ) to the description section of the repo (click the "Edit" button and fill in the website address then click "Save").

3. Add the link of the online vignette to the `README.md` so users know where to look to get started.

https://github.com/openjournals/joss-reviews/issues/1077
