### 2018-11-12 sumbose/iRF#17 @road2stat
[JOSS review] Add DOI to references
Please add the corresponding DOI to the references (if available) in `paper.bib`. [Format example here](https://github.com/terrytangyuan/dml/blob/master/paper/paper.bib).

https://github.com/openjournals/joss-reviews/issues/1077
