### 2018-11-13 a6e8823
reformat affiliations
