+++ paper.md
@@ -6,9 +6,9 @@ tags:
   - Interpretable machine learning
 authors:
   - name: Sumanta Basu
-    affiliation: "&#8727;, 1, 2"
+    affiliation: a, 1, 2"
   - name: Karl Kumbier
-    affiliation: "&#8727;, 3"
+    affiliation: a, 3"
   - name: James B. Brown
     affiliation: "3, 4, 5"
   - name: Bin Yu
@@ -27,7 +27,7 @@ affiliations:
  - name: Department of Electrical Engineering and Computer Sciences, University of California, Berkeley 
    index: 6
  - name: Denotes equal contribution
-   index: &#8727;;
+   index: a;
 date: 2 October 2018
 bibliography: paper.bib
 ---
