### 2018-11-13 22191a8
reformat affiliations
