### 2018-11-13 59749e2
update readme with vignette
