### 2018-11-13 3928df5
helper scripts for signed iRF
