+++ paper.bib
@@ -17,7 +17,8 @@
   number={1},
   pages={5--32},
   year={2001},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1023/A:1010933404324}
 }
 
 @article{amaratunga2008enriched,
@@ -28,7 +29,8 @@
   number={18},
   pages={2010--2014},
   year={2008},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btn356}
 }
 
 @article{liaw2002classification,
@@ -47,5 +49,15 @@
   journal={Proceedings of the National Academy of Sciences},
   pages={201711236},
   year={2018},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.1711236115}
+}
+
+@article{kumbier2018refining,
+  title={Refining interaction search trhough signed iterative Random Forests},
+  author={Kumbier, Karl and Basu, Sumanta and Brown, James B and Celniker,
+    Susan and Yu, Bin},
+  journal={arXiv preprint},
+  publisher={arXiv:1810.07287},
+  year={2018}
 }
