+++ paper.bib
@@ -57,7 +57,6 @@
   title={Refining interaction search trhough signed iterative Random Forests},
   author={Kumbier, Karl and Basu, Sumanta and Brown, James B and Celniker,
     Susan and Yu, Bin},
-  journal={arXiv preprint},
-  publisher={arXiv:1810.07287},
+  journal={arXiv:1810.07287},
   year={2018}
 }
