+++ R/zzz.R
@@ -1,5 +1,0 @@
-.onAttach <- function(libname, pkgname) {
-    RFver <- read.dcf(file=system.file("DESCRIPTION", package=pkgname),
-                      fields="Version")
-    packageStartupMessage(paste(pkgname, RFver))
-}

[…]
