### 2018-11-14 sumbose/iRF#12 @nanxstats
-ok great, it looks this has been fixed, whedon is generating pdf properly. Closing now...
