### 2018-11-14 sumbose/iRF#13 @nanxstats
Looks like the above items are now fixed in the readme. Good job! Closing for now.
