### 2018-11-14 sumbose/iRF#14 @nanxstats
The HTML vignette is now linked in the readme so I would consider this fixed.
