### 2018-11-14 sumbose/iRF#17 @nanxstats
Hi @karlkumbier -I saw most of the papers are now with a DOI. Would you mind also adding the DOI for the preprint "Refining interaction search trhough signed iterative Random Forests" ("trhough" should be "through")? I guess bioRxiv preprints all have one.
