### 2018-11-14 sumbose/iRF#18 @road2stat
[JOSS review] Fix GitHub version installation issue
Currently, when running `devtools::install_github("sumbose/iRF")`, I'm seeing

```
** testing if installed package can be loaded
Error: package or namespace load failed for ‘iRF’ in namespaceExport(ns, exports):
 undefined exports: interactPredict
Error: loading failed
Execution halted
```

Should be a quick fix.

https://github.com/openjournals/joss-reviews/issues/1077
