### 2018-11-14 sumbose/iRF#19 @road2stat
[JOSS review] Fix example code in vignette
Currently, when using the **CRAN version** and running this code block in [the vignette](https://www.stat.berkeley.edu/~kkumbier/vignette.html):

```r
> rforest$node.feature[1:10, c(1:5, p + 1:5)]
```

it shows:

```
Error in intI(j, n = d[2], dn[[2]], give.dn = FALSE) :
  index larger than maximal 250
```

and almost everything after this also breaks.

https://github.com/openjournals/joss-reviews/issues/1077
