### 2018-11-14 sumbose/iRF#20 @road2stat
[JOSS review] Use detectCores() instead of specific number in example
Currently, the code in the vignette uses `n.cores <- 4`. I'd recommend replacing it with the more flexible `n.cores <- detectCores()` so that users with different numbers of cpu cores can achieve maximal parallelization on their machines.

https://github.com/openjournals/joss-reviews/issues/1077
