### 2018-11-14 fdade44
update namespace and change default cores
