+++ paper.bib
@@ -57,6 +57,7 @@
   title={Refining interaction search trhough signed iterative Random Forests},
   author={Kumbier, Karl and Basu, Sumanta and Brown, James B and Celniker,
     Susan and Yu, Bin},
-  journal={arXiv:1810.07287},
+  journal={bioRxiv},
+  doi={10.1101/467498},
   year={2018}
 }
