+++ R/gRIT.R
@@ -11,10 +11,12 @@ gRIT <- function(x, y,
                          class.cut=NULL),
                  signed=TRUE,
                  ints.full=NULL,
-                 n.core=1) {
+                 n.core=-1) {
 
   out <- list()
   class.irf <- is.factor(y)
+  if (ncore == -1) n.core <- detectCores()  
+  registerDoMC(n.core)
 
   # Check rit parameters and set default values if not specified
   if (is.null(rand.forest) & is.null(read.forest))
@@ -94,12 +96,12 @@ gRIT <- function(x, y,
      
       ints.sub <- lapply(ints.full, intSubsets)
       ints.sub <- unique(unlist(ints.sub, recursive=FALSE))
-      imp <- mclapply(ints.sub, intImportance,
-                      nf=read.forest$node.feature,
-                      yprec=yprec, 
-                      select.id=idcl, 
-                      weight=ndcnt,
-                      mc.cores=n.core)
+      suppressWarnings(
+      imp <- foreach(int=ints.sub) %dorng% {
+        intImportance(int, nf=read.forest$node.feature,
+                      yprec=yprec, select.id=idcl, 
+                      weight=ndcnt)
+      })
       out$imp <- rbindlist(imp) 
       imp.test <- lapply(ints.full, subsetTest, importance=out$imp, ints=ints.sub)
       imp.test <- rbindlist(imp.test)
