### 2018-11-14 a7346a6
cleaning for CRAN
