+++ R/surfacePlot.R
@@ -1,0 +1,261 @@
+plotInt2 <- function(rectangles, interact, x, y, 
+                     grid.size=100, col.pal=c('blue', 'yellow'), 
+                     xlab=NULL, ylab=NULL, zlab=NULL,
+                     range.color=NULL,
+                     z.range=c(0,1),
+                     n.cols=100, grids=NULL, axes=TRUE,
+                     varnames.grp=colnames(x),
+                     pred.prob=FALSE, min.node=10) {
+  # Generates surface map of order-2 interaction
+  # args:
+  #   rectangles: hyprerectangle list as generated by forestHyperrectangle
+  #   interact.plot: order-2 interaction to plot
+  #   grid.size: number of bins to plot surface map over
+  #   col.pal: color palette of surface map
+  #   xlab, ylab, zlab: axis labels
+  #   range.color: range of color values
+  #   z.range: range for response axis
+  #   n.cols: number of colors in color pal
+  #   grids: user generated grid to plot over. If supplied, grid.size is 
+  #     ignored
+  #   axes: T/F indicating whether axes should be plotted
+  #   varnames.group: character vector indicating feature grouping
+  #   groupFUN: function to use for grouping columns of x
+  #   pred.prob: T/F indicating whether the z axis should indicate predicted
+  #     probability from the forest or raw data distribution
+  #   min.node.size: minimum leaf node size to use for grid
+  require(rgl)
+  
+  if (is.null(varnames.grp)) varnames.grp <- 1:ncol(x)
+  interact <- int2Id(interact, varnames.grp, signed=TRUE, split=TRUE)
+  stopifnot(length(interact) == 2) 
+  
+  n <- nrow(x)
+  p <- ncol(x)
+  if (is.factor(y)) y <- as.numeric(y) - 1
+  
+  # generate grid to plot surface over either as raw values or quantiles
+  int1 <- interact[1]
+  int2 <- interact[2]
+  
+  if (is.null(grids)) {
+    i1 <- int1 %% p + p * (int1 %% p == 0)
+    i2 <- int2 %% p + p * (int2 %% p == 0)
+    g1 <- seq(min(x[,i1]), max(x[,i1]), length.out=grid.size)
+    g2 <- seq(min(x[,i2]), max(x[,i2]), length.out=grid.size)
+    g1n <- round(g1, 2)
+    g2n <- round(g2, 2)
+  } else {
+    g1 <- grids$g1
+    g2 <- grids$g2
+    g1n <- seq(0, 1, length.out=length(g1))
+    g2n <- seq(0, 1, length.out=length(g2))
+    grid.size <- length(g1)
+  }
+  
+  # Evaluate responses over grid based on RF hyperrectangles
+  grid <- matrix(0, nrow=grid.size, ncol=grid.size)
+  rectangles <- filter(rectangles, size.node >= min.node)
+  removed <- rep(FALSE, nrow(rectangles))
+  for (i in 1:nrow(rectangles)) {
+    
+    rtemp <- rectangles[i,]
+    tt1 <- unlist(rtemp$splits)[unlist(rtemp$vars) == int1]
+    tt2 <- unlist(rtemp$splits)[unlist(rtemp$vars) == int2]
+    
+    if (int1 <= p) {
+      idcs1 <- g1 < tt1
+      x1 <- x[, intp1] < tt1
+    } else {
+      idcs1 <- g1 >= tt1
+      int11 <- int1 %% p + p * (int1 %% p == 0)
+      x1 <- x[,int11] >= tt1
+    }
+    
+    if (int2 <= p) {
+      idcs2 <- g2 < tt2
+      x2 <- x[, intp2] < tt2
+    } else {
+      idcs2 <- g2 >= tt2
[…]
