### 2018-11-15 sumbose/iRF#18 @nanxstats
Now the GitHub version compiles. Closing...
