### 2018-11-15 sumbose/iRF#20 @nanxstats
This is confirmed to be fixed in https://github.com/sumbose/iRF/commit/fdade440c10ce057de4ee0fcb56255c57cd6f6c6.
