### 2018-11-15 0c787ec
update license and man files for CRAN
