+++ R/readForest.R
@@ -11,7 +11,7 @@ readForest <- function(rfobj, x,
   if (is.null(varnames.grp)) varnames.grp <- 1:ncol(x)
  
   if (n.core == -1) n.core <- detectCores()
-  registerDoMC(n.core)
+  registerDoParallel(n.core)
   ntree <- rfobj$ntree
   n <- nrow(x)
   p <- length(unique(varnames.grp))
