+++ man/importance.Rd
@@ -47,12 +47,11 @@
   \code{\link{randomForest}}, \code{\link{varImpPlot}}
 }
 \examples{
-set.seed(4543)
-data(mtcars)
-mtcars.rf <- randomForest(mpg ~ ., data=mtcars, ntree=1000,
-                          keep.forest=FALSE, importance=TRUE)
-importance(mtcars.rf)
-importance(mtcars.rf, type=1)
+set.seed(454)
+iris.rf <- randomForest(Species ~ ., data=iris, ntree=1000,
+                        keep.forest=FALSE, importance=TRUE)
+importance(iris.rf)
+importance(iris.rf, type=1)
 }
 %\author{}
 \keyword{regression}
