+++ man/varImpPlot.Rd
@@ -29,10 +29,9 @@ varImpPlot(x, sort=TRUE, n.var=min(30, nrow(x$importance)),
 }
 \examples{
 set.seed(4543)
-data(mtcars)
-mtcars.rf <- randomForest(mpg ~ ., data=mtcars, ntree=1000, keep.forest=FALSE,
-                          importance=TRUE)
-varImpPlot(mtcars.rf)
+iris.rf <- randomForest(Species ~ ., data=iris, ntree=1000, keep.forest=FALSE,
+                        importance=TRUE)
+varImpPlot(iris.rf)
 }
 \author{Andy Liaw \email{andy\_liaw@merck.com}}
 \keyword{regression}
