+++ man/imports85.Rd
@@ -52,13 +52,6 @@ data(imports85)
 data(imports85)
 imp85 <- imports85[,-2]  # Too many NAs in normalizedLosses.
 imp85 <- imp85[complete.cases(imp85), ]
-## Drop empty levels for factors.
-imp85[] <- lapply(imp85, function(x) if (is.factor(x)) x[, drop=TRUE] else x)
-
-price.rf <- randomForest(price ~ ., imp85, do.trace=10, ntree=100)
-print(price.rf)
-numDoors.rf <- randomForest(numOfDoors ~ ., imp85, do.trace=10, ntree=100)
-print(numDoors.rf)
 }
 \author{Andy Liaw}
 \keyword{datasets}
