+++ man/randomForest.Rd
@@ -212,7 +212,7 @@
   5-32.
 
   Breiman, L (2002), ``Manual On Setting Up, Using, And Understanding
-  Random Forests V3.1'', \url{http://oz.berkeley.edu/users/breiman/Using_random_forests_V3.1.pdf}.
+  Random Forests V3.1'' 
 }
 \author{
   Sumanta Basu \email{sumbose@berkeley.edu}, based on source codes from
@@ -239,46 +239,8 @@ pairs(cbind(iris[,1:4], iris.mds$points), cex=0.6, gap=0,
 par(op)
 print(iris.mds$GOF)
 
-## The `unsupervised' case:
-set.seed(17)
-iris.urf <- randomForest(iris[, -5])
-MDSplot(iris.urf, iris$Species)
-
-## stratified sampling: draw 20, 30, and 20 of the species to grow each tree.
-(iris.rf2 <- randomForest(iris[1:4], iris$Species, 
-                          sampsize=c(20, 30, 20)))
-
-## Regression:
-## data(airquality)
-set.seed(131)
-ozone.rf <- randomForest(Ozone ~ ., data=airquality, mtry=3,
-                         importance=TRUE, na.action=na.omit)
-print(ozone.rf)
-## Show "importance" of variables: higher value mean more important:
-round(importance(ozone.rf), 2)
-
-## "x" can be a matrix instead of a data frame:
-set.seed(17)
-x <- matrix(runif(5e2), 100)
-y <- gl(2, 50)
-(myrf <- randomForest(x, y))
-(predict(myrf, x))
-
-## "complicated" formula:
-(swiss.rf <- randomForest(sqrt(Fertility) ~ . - Catholic + I(Catholic < 50),
-                          data=swiss))
-(predict(swiss.rf, swiss))
-## Test use of 32-level factor as a predictor:
-set.seed(1)
-x <- data.frame(x1=gl(53, 10), x2=runif(530), y=rnorm(530))
-(rf1 <- randomForest(x[-3], x[[3]], ntree=10))
-
 ## Grow no more than 4 nodes per tree:
 (treesize(randomForest(Species ~ ., data=iris, maxnodes=4, ntree=30)))
-
-## test proximity in regression
-iris.rrf <- randomForest(iris[-1], iris[[1]], ntree=101, proximity=TRUE, oob.prox=FALSE)
-str(iris.rrf$proximity)
 }
 \keyword{classif}% at least one, from doc/KEYWORDS
 \keyword{regression}
