### 2018-11-15 sumbose/iRF#19 @nanxstats
The example code now runs in `vignettes/vignette.Rmd`. @karlkumbier -just one small request, please also remember to update the compiled HTML linked in the README to reflect this change. I'm going to close this now.
