### 2018-11-15 #1077 @road2stat
@karlkumbier and @arokem - the software and the paper have passed my review. It can be a very useful tool for making random forests interpretable - great job!
