### 2018-11-20 81839ba
remove track nodes
