+++ R/randomForest.default.R
@@ -15,7 +15,7 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
            proximity, oob.prox=proximity,
            norm.votes=TRUE, do.trace=FALSE,
            keep.forest=!is.null(y) && is.null(xtest), corr.bias=FALSE,
-           keep.inbag=FALSE, track.nodes=FALSE,...) {
+           keep.inbag=FALSE, ...) {
     
     if (length(mtry.select.prob)!=ncol(x))
       stop('length of mtry.select.prob != ncol(x)')
@@ -241,17 +241,11 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
     
     # pre allocate matrices for tracking feature and observation data at each
     # node
-    if (track.nodes) {
-      obs.nodes <- matrix(0L, nrow=n, ncol=ntree)
-    } else {
-      obs.nodes <- 0L
-    }
-    
+   
     if (classRF) {
       cwt <- classwt
       threshold <- cutoff
       error.test <- if (labelts) double((nclass+1) * ntree) else double(1)
-     #46 
       rfout <- .C("classRF",
                   x = x,
                   xdim = as.integer(c(p, n)),
@@ -274,8 +268,6 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
                   ntree = as.integer(ntree),
                   mtry = as.integer(mtry),
                   selprob = as.double(mtry.select.prob),
-                  obsnodes = obs.nodes,
-                  tracknodes = as.integer(track.nodes),
                   subsetVar = subsetVar, 
                   subsetVarCard = subsetVarCard, 
                   maxCard = maxCard,
@@ -308,6 +300,7 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
                   errts = error.test,
                   inbag = if (keep.inbag)
                     matrix(integer(n * ntree), n) else integer(n),
+                  DUP=FALSE,
                   PACKAGE="iRF"
       )[-1]  
       if (keep.forest) {
@@ -394,7 +387,6 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
                          nrnodes = max.nodes, ntree = ntree,
                          nclass = nclass, xlevels=xlevels)
                   },
-                  obs.nodes = if (!track.nodes) NULL else rfout$obsnodes,
                   test = if(!testdat) NULL else list(
                     predicted = out.class.ts,
                     err.rate = if (labelts) t(matrix(rfout$errts, nclass+1,
@@ -405,8 +397,7 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
                     proximity = if(proximity) matrix(rfout$proxts, nrow=ntest,
                                                      dimnames = list(xts.row.names, c(xts.row.names,
                                                                                       x.row.names))) else NULL),
-                  inbag = if (keep.inbag) rfout$inbag else NULL,
-                  y = as.factor(y))
+                  inbag = if (keep.inbag) rfout$inbag else NULL)
     }
     
     else {
@@ -414,64 +405,55 @@ mylevels <- function(x) if (is.factor(x)) levels(x) else 0
       y <- y - ymean
       ytest <- ytest - ymean
       
-      # pre allocate matrices for tracking feature and observation data at
-      # each
-      # node
-      if (track.nodes) {
-        obs.nodes <- matrix(0L, nrow=n, ncol=ntree)
-      } else {
-        obs.nodes <- 0L
-      }
       rfout <- .C("regRF",
-                  x,
-                  as.double(y),
-                  as.integer(c(n, p)),
-                  as.integer(sampsize),
-                  as.integer(nodesize),
-                  as.integer(nrnodes),
-                  as.integer(ntree),
-                  as.integer(mtry),
[…]
