### 2018-11-21 659fb08
clean namespace and remove FSInteract dependency
