+++ NAMESPACE
@@ -17,7 +17,7 @@ importFrom(stats, delete.response, mad, median, model.frame,
            var, weighted.mean)
 importFrom(doParallel, registerDoParallel)
 importFrom(doRNG, "%dorng%")
-importFrom(foreach, foreach)
+importFrom(foreach, foreach, "%dopar%")
 importFrom(parallel, detectCores)
 
 export(combine, getTree, grow, importance, margin, MDSplot, na.roughfix,
