+++ DESCRIPTION
@@ -14,7 +14,6 @@ Imports:
     methods,
     doMC,
     doRNG,
-    FSInteract,
     foreach,
     stringr
 Suggests:
