+++ inst/doc/vignette.Rmd
@@ -1,0 +1,268 @@
+## iterative Random Forests (iRF)
+
+The R package `iRF` implements iterative Random Forests, a method for
+iteratively growing ensemble of weighted decision trees, and detecting
+high-order feature interactions by analyzing feature usage on decision paths.
+This version uses source codes from the R package `randomForest` by Andy Liaw
+and Matthew Weiner, the original Fortran codes by Leo Breiman and Adele Cutler,
+and source codes from the R package `FSInteract` by Hyun Jik Kim and Rajen D.
+Shah.
+
+`iRF` can be loaded using the `library` command:
+```{r, message=FALSE}
+library(iRF)
+```
+
+`iRF` adds two new features to Breiman's original Random Forest workflow:
+
+##  A. Weighted random forest
+
+Unlike Breiman's original random forest, which uses uniform random sampling to
+select \texttt{mtry} variables during each node split, the 'randomForest'
+function in 'iRF' allows non-uniform sampling using a given vector of
+nonnegative weights (e.g., feature importances from a previous model fit). In
+particular, given a vector of weights $\mathbf{w} = (w_1, \ldots, w_p)$, the
+`mtry` variables are selected so that $P(\text{variable }j \text{ is selected})
+= w_j/\sum_{j=1}^p w_j$, for $j = 1, \ldots, p$. 
+
+Based on this weighting scheme, we can iteratively grow weighted random forests,
+where Gini importances from the previous random forest fit are used as weights
+in the current iteration.
+
+```{r seed_random}
+# set seed for random number generation
+set.seed(47)                           
+```
+
+### Binary Classification:
+We simulate some data for a binary classification exercise. Out of $250$
+features, only three $\{1, 2, 3\}$ are important.  
+```{r simulate_data}
+  # simulate data for classification
+  n <- 500
+  p <- 250
+  X <- matrix(rnorm(n * p), nrow=n)
+
+  Y <- (X[,1] > 0 & X[,2] > 0 & X[,3] > 0)
+  Y <- as.factor(as.numeric(Y))
+
+  train.id <- 1:(n / 2)
+  test.id <- setdiff(1:n, train.id)
+```
+
+Next, we iteratively grow weighted random forests and use feature importances
+from last iteration as weights.  
+```{r fit_irf}
+sel.prob <- rep(1/p, p)
+
+# iteratively grow RF, use Gini importance of features as weights
+rf <- list()
+for (iter in 1:4){
+  rf[[iter]] <- randomForest(x=X[train.id,], y=Y[train.id], 
+                             xtest=X[test.id,], ytest=Y[test.id], 
+                             mtry.select.prob=sel.prob)
+
+  # update selection probabilities for next iteration
+  sel.prob <- rf[[iter]]$importance
+}
+```
+
+### ROC curve
+We can measure performance of different iterations of RF on the test set:
+```{r print_auc, fig.height = 6, fig.width = 6, message=FALSE}
+library(AUC)
+plot(0:1, 0:1, type='l', lty = 2, xlab = 'FPR', ylab = 'TPR', main='ROC Curve')
+for (iter in 1:4){
+  # performance on test set
+  cat(paste('iter = ', iter, ':: '))
+  roc.info <- roc(rf[[iter]]$test$votes[,2], Y[test.id])
+  lines(roc.info$fpr, roc.info$tpr, type='l', col=iter, lwd=2)
+  cat(paste('AUROC: ', round(100*auc(roc.info), 2), '%\n', sep=''))
+} 
+legend('bottomright', legend=paste('iter:', 1:iter), col=1:iter, lwd=2, bty='n')
+```
+
+### Variable Importance
+The outputs  of `iRF::randomForest` are objects of class `randomForest` and can
[…]
