### 2018-11-29 #1077 @arokem
Sorry for my slowness here...

@road2stat : thank you for the review!

@karlkumbier : I believe that your paper is ready to be accepted for publication in JOSS.

Could you please create an archive of the accepted version of the software (e.g., using [zenodo](https://zenodo.org)) and post the DOI here?
