### 2018-12-05 #1077 @danielskatz
👋 @karlkumbier -  your paper is now accepted into JOSS ⚡️🚀💥
